python/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "maptitude"
version = "0.1.6"
description = "High-performance crystallographic electron density computation and scoring"
requires-python = ">=3.10"
dependencies = ["openeye-toolkits>=2025.2"]

[tool.setuptools]
packages = ["maptitude"]
